+++ python/featomic/featomic/clebsch_gordan/__init__.py
@@ -3,4 +3,7 @@ from ._cg_product import ClebschGordanProduct  # noqa: F401
 from ._coefficients import calculate_cg_coefficients  # noqa: F401
 from ._density_correlations import DensityCorrelations  # noqa: F401
 from ._equivariant_power_spectrum import EquivariantPowerSpectrum  # noqa: F401
+from ._equivariant_power_spectrum_by_pair import (  # noqa: F401
+    EquivariantPowerSpectrumByPair,  # noqa: F401
+)  # noqa: F401
 from ._power_spectrum import PowerSpectrum  # noqa: F401